+++ serverless/src/dynamo-scan-victims.js
