+++ serverless/src/dynamo-write-victims.js
